fix(coding-agent): redact hook GitHub tokens
