+++ packages/coding-agent/test/suite/hooks-command.test.ts
@@ -1,30 +1,20 @@
-import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
-import { join, resolve } from "node:path";
+import { existsSync, readFileSync, writeFileSync } from "node:fs";
 import { afterEach, describe, expect, it } from "vitest";
-import hooksExtension from "../../src/core/extensions/builtin/hooks/index.ts";
-import type { ExtensionUIContext } from "../../src/core/extensions/index.ts";
-import type { LoadedHookSources } from "../../src/core/extensions/types.ts";
-import type { ResourceLoader } from "../../src/core/resource-loader.ts";
-import { theme } from "../../src/modes/interactive/theme/theme.ts";
-import { createTestExtensionsResult, createTestResourceLoader } from "../utilities.ts";
-import { createHarness, type Harness } from "./harness.ts";
-
-type Notification = {
-	readonly message: string;
-	readonly type: "info" | "warning" | "error" | undefined;
-};
+import {
+	cleanupHooksCommandHarnesses,
+	createHooksCommandHarness,
+	firstHookId,
+	lastNotification,
+	readState,
+} from "./hooks-command-harness.ts";
 
 const USAGE = "Usage: /hooks [list|diagnostics|trust <id>|disable <id>|enable <id>|reload]";
 const SECRET = "sk-test-secret-should-not-render";
 const API_KEY_SECRET = "plain-api-key-value-should-not-render";
 const PASSWORD_SECRET = "plain-password-value-should-not-render";
 
-const harnesses: Harness[] = [];
-
 afterEach(() => {
-	while (harnesses.length > 0) {
-		harnesses.pop()?.cleanup();
-	}
+	cleanupHooksCommandHarnesses();
 });
 
 describe("builtin hooks /hooks command", () => {
@@ -248,132 +238,3 @@ describe("builtin hooks /hooks command", () => {
 		await expect(command?.getArgumentCompletions?.("trust ")).resolves.toEqual([]);
 	});
 });
-
-async function createHooksCommandHarness(options: { readonly projectHooks: unknown }) {
-	const notifications: Notification[] = [];
-	const payloads: unknown[] = [];
-	const extensionsResult = await createTestExtensionsResult([{ factory: hooksExtension, path: "<builtin:hooks>" }]);
-	const baseResourceLoader = createTestResourceLoader({ extensionsResult });
-	let reloadCalls = 0;
-	let hookSources: LoadedHookSources | undefined;
-	const resourceLoader: ResourceLoader = {
-		...baseResourceLoader,
-		getLoadedHookSources: () => {
-			if (hookSources === undefined) {
-				throw new Error("hook sources not initialized");
-			}
-			return hookSources;
-		},
-		reload: async () => {
-			reloadCalls += 1;
-		},
-	};
-	const harness = await createHarness({
-		resourceLoader,
-		withConfiguredAuth: false,
-		onPayload: (payload) => payloads.push(payload),
-	});
-	harnesses.push(harness);
-	const agentDir = join(harness.tempDir, "agent");
-	const projectHooksPath = join(harness.tempDir, ".senpi", "hooks.json");
-	const projectStatePath = join(harness.tempDir, ".senpi", "hooks-state.json");
-	mkdirSync(agentDir, { recursive: true });
-	mkdirSync(join(harness.tempDir, ".senpi"), { recursive: true });
-	writeFileSync(projectHooksPath, JSON.stringify(options.projectHooks), "utf-8");
-	hookSources = {
-		agentDir,
-		cwd: harness.tempDir,
-		globalHookSourcePaths: [],
-		globalHooksPath: join(agentDir, "hooks.json"),
-		preSessionHookSourcePaths: [],
-		projectHookSourcePaths: [],
-		projectHooksPath,
-		runtimeHookSourcePaths: [],
-	};
-	await harness.session.bindExtensions({
-		uiContext: createUiContext((message, type) => notifications.push({ message, type })),
-		commandContextActions: {
-			waitForIdle: async () => {},
-			newSession: async () => ({ cancelled: false }),
-			fork: async () => ({ cancelled: false }),
-			navigateTree: async () => ({ cancelled: false }),
-			switchSession: async () => ({ cancelled: false }),
-			reload: async () => {
-				await resourceLoader.reload();
-			},
-		},
-	});
-	return {
-		harness,
-		notifications,
-		payloads,
-		projectHooksPath,
-		projectStatePath,
-		get reloadCalls() {
-			return reloadCalls;
-		},
-	};
-}
-
-function createUiContext(
-	onNotify: (message: string, type: "info" | "warning" | "error" | undefined) => void,
-): ExtensionUIContext {
-	return {
-		select: async () => undefined,
-		confirm: async () => false,
-		input: async () => undefined,
-		notify: onNotify,
-		onTerminalInput: () => () => {},
-		setStatus: () => {},
-		setWorkingMessage: () => {},
-		setWorkingVisible: () => {},
-		setWorkingIndicator: () => {},
-		setHiddenThinkingLabel: () => {},
-		setWidget: () => {},
-		setFooter: () => {},
-		setHeader: () => {},
-		setTitle: () => {},
-		custom: async <T>() => undefined as T,
-		pasteToEditor: () => {},
-		setEditorText: () => {},
-		getEditorText: () => "",
-		editor: async () => undefined,
-		addAutocompleteProvider: () => {},
-		setEditorComponent: () => {},
-		getEditorComponent: () => undefined,
-		get theme() {
-			return theme;
-		},
-		getAllThemes: () => [],
-		getTheme: () => undefined,
-		setTheme: () => ({ success: false, error: "themes are not used by /hooks tests" }),
-		getToolsExpanded: () => false,
-		setToolsExpanded: () => {},
-	};
-}
-
-function lastNotification(notifications: readonly Notification[]): Notification {
-	const notification = notifications.at(-1);
-	if (notification === undefined) {
-		throw new Error("expected notification");
-	}
-	return notification;
-}
-
-function firstHookId(message: string): string {
-	const match = /^- (hk_[^ ]+)/m.exec(message);
-	if (match === null) {
-		throw new Error(`expected hook ID in message:\n${message}`);
-	}
-	return match[1];
-}
-
-function readState(path: string): {
-	readonly hooks: Record<string, { readonly enabled: boolean; readonly trustedHash?: string }>;
-} {
-	const parsed = JSON.parse(readFileSync(resolve(path), "utf-8"));
-	if (typeof parsed !== "object" || parsed === null || !("hooks" in parsed)) {
-		throw new Error("invalid hooks state");
-	}
-	return parsed as { readonly hooks: Record<string, { readonly enabled: boolean; readonly trustedHash?: string }> };
-}
